+++ navigator-compose/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version `0.1-alpha19` _(2021-09-01)_
+
+- **Added**: `handleOnDismissRequest` API in `CreateDialog` for intercepting dismiss request.
+- **Fixed**: Incorrect of popping dialog on onBackPressed().
+
 ## Version `0.1-alpha18` _(2021-08-31)_
 
 - **Breaking Change**: Renamed `setDefaultBackPressEnabled(boolean)` to `disableDefaultBackPressLogic()`.